requirements.txt
alembic==1.10.2
anyio==3.6.2
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==1.7.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
fastapi-pagination==0.11.4
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
Mako==1.2.4
manage-fastapi==1.1.1
MarkupSafe==2.1.2
passlib==1.7.4
poyo==0.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pyasn1==0.5.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
questionary==1.10.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
text-unidecode==1.3
typer==0.4.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
